requirements.txt
numpy
pandas
torch
torchmetrics
boto3
pyathena
python-dotenv